+++ packages/remirror__extension-emoji/src/emoji-extension.ts
@@ -84,7 +84,7 @@ export class EmojiExtension extends NodeExtension<EmojiOptions> {
       inline: true,
 
       atom: true,
-      attrs: { ...extra.defaults(), code: {} },
+      attrs: { ...extra.defaults(), code: { validate: 'string|null' } },
       parseDOM: [
         {
           tag: `span[${EMOJI_DATA_ATTRIBUTE}`,