+++ packages/remirror__extension-entity-reference/src/entity-reference-extension.ts
@@ -22,7 +22,7 @@ import {
   within,
 } from '@remirror/core';
 import type { CreateEventHandlers } from '@remirror/extension-events';
-import { Node } from '@remirror/pm/model';
+import { Node, TagParseRule } from '@remirror/pm/model';
 import { EditorStateConfig, TextSelection } from '@remirror/pm/state';
 import { DecorationSet } from '@remirror/pm/view';
 
@@ -86,7 +86,7 @@ export class EntityReferenceExtension extends MarkExtension<EntityReferenceOptio
       excludes: '',
       attrs: {
         ...extra.defaults(),
-        id: { default: '' },
+        id: { default: '', validate: 'string' },
       },
       toDOM: (mark: Mark) => [
         'span',
@@ -107,7 +107,7 @@ export class EntityReferenceExtension extends MarkExtension<EntityReferenceOptio
             const id = node.getAttribute(idAttr);
             return { ...extra.parse(node), id };
           },
-        },
+        } satisfies TagParseRule,
         ...(override.parseDOM ?? []),
       ],
     };