+++ packages/remirror__extension-tables/src/table-utils.ts
@@ -86,7 +86,7 @@ function setCellAttrs(node: ProsemirrorNode) {
   }
 
   if (node.attrs.background) {
-    attrs.style = `${attrs.style ?? ''}background-color: ${node.attrs.background as string};`;
+    attrs.style = `${node.attrs.style ?? ''}background-color: ${node.attrs.background as string};`;
     attrs['data-background-color'] = node.attrs.background;
   }
 
@@ -102,10 +102,10 @@ export function createTableNodeSchema(
 ): Record<'table' | 'tableRow' | 'tableCell' | 'tableHeaderCell', TableSchemaSpec> {
   const cellAttrs = {
     ...extra.defaults(),
-    colspan: { default: 1 },
-    rowspan: { default: 1 },
-    colwidth: { default: null },
-    background: { default: null },
+    colspan: { default: 1, validate: 'number' },
+    rowspan: { default: 1, validate: 'number' },
+    colwidth: { default: null, validate: 'number|null' },
+    background: { default: null, validate: 'string|null' },
   };
 
   return {