+++ packages/prosemirror-suggest/CHANGELOG.md
@@ -49,6 +49,17 @@
   - @remirror/core-helpers@4.0.0-beta.0
   - @remirror/types@2.0.0-beta.0
 
+## 2.0.7
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- 7caff8388: Bump ProseMirror to latest versions to address potential XSS vulnerability found in ProseMirror's DOMSerializer
+
+  See: https://discuss.prosemirror.net/t/heads-up-xss-risk-in-domserializer/6572
+
 ## 2.0.6
 
 > 2023-07-31