+++ packages/prosemirror-trailing-node/CHANGELOG.md
@@ -45,6 +45,25 @@
   - @remirror/core-constants@3.0.0-beta.0
   - @remirror/core-helpers@4.0.0-beta.0
 
+## 2.0.9
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- 7caff8388: Bump ProseMirror to latest versions to address potential XSS vulnerability found in ProseMirror's DOMSerializer
+
+  See: https://discuss.prosemirror.net/t/heads-up-xss-risk-in-domserializer/6572
+
+## 2.0.8
+
+> 2024-03-05
+
+### Patch Changes
+
+- 45886720e: Remove dependency of `prosemirror-trailing-node` on `@remirror/core-helpers`.
+
 ## 2.0.7
 
 > 2023-07-31