+++ packages/remirror/CHANGELOG.md
@@ -782,6 +782,26 @@
   - @remirror/dom@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 2.0.39
+
+> 2024-02-21
+
+### Patch Changes
+
+- 00a41e38f: Fix the html sanitizing on HTML output from markdown. Removed the default html sanitizer because it doesn't provide any security guarantees and it's not been called due to a bug in the markdown extension.
+- Updated dependencies [00a41e38f]
+  - @remirror/extension-markdown@2.0.14
+
+## 2.0.38
+
+> 2023-12-12
+
+### Patch Changes
+
+- 55602747f: Fix a bug where calling `commands.setTableCellBackground` overrides the style attribute of the table cell node.
+- Updated dependencies [55602747f]
+  - @remirror/extension-tables@2.4.1
+
 ## 2.0.37
 
 > 2023-11-12