+++ packages/remirror__extension-code-block/CHANGELOG.md
@@ -102,6 +102,17 @@
   - @remirror/theme@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 2.0.19
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- Updated dependencies [7caff8388]
+- Updated dependencies [7caff8388]
+  - @remirror/pm@2.0.9
+
 ## 2.0.18
 
 > 2023-08-03