+++ packages/remirror__extension-codemirror5/CHANGELOG.md
@@ -96,6 +96,17 @@
   - @remirror/messages@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 2.0.14
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- Updated dependencies [7caff8388]
+- Updated dependencies [7caff8388]
+  - @remirror/pm@2.0.9
+
 ## 2.0.13
 
 > 2023-04-26