+++ packages/remirror__extension-image/CHANGELOG.md
@@ -81,6 +81,18 @@
   - @remirror/theme@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 2.1.11
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- Updated dependencies [7caff8388]
+- Updated dependencies [7caff8388]
+  - prosemirror-resizable-view@2.0.15
+  - @remirror/pm@2.0.9
+
 ## 2.1.10
 
 > 2023-07-31