+++ packages/remirror__extension-yjs/CHANGELOG.md
@@ -75,6 +75,21 @@
   - @remirror/messages@3.0.0-beta.0
   - @remirror/pm@3.0.0-beta.0
 
+## 3.0.16
+
+> 2024-07-17
+
+### Patch Changes
+
+- 7caff8388: Add a validate property to each of the Node or Mark attributes used in Remirror
+- 7caff8388: Bump ProseMirror to latest versions to address potential XSS vulnerability found in ProseMirror's DOMSerializer
+
+  See: https://discuss.prosemirror.net/t/heads-up-xss-risk-in-domserializer/6572
+
+- Updated dependencies [7caff8388]
+- Updated dependencies [7caff8388]
+  - @remirror/pm@2.0.9
+
 ## 3.0.15
 
 > 2023-07-31